Sets, Lists, Maps
Tried to use a TreeMap with a Comparator, but while a Map can contain duplicate values with unique keys this removed my employee with same last names (lastName was my comparator value)  I used employee as the key e1-e4 and ssn as a value.  I had to do this because the Comparator wanted an Object and String could not be cast to Employee - of course.

Not sure how to proceed...Could use a list to hold the map and then sort the list as a collection???